functions/cookie_authorizer/requirements.txt
PyJWT==2.6.0
cryptography==38.0.4
